Wifi: 8852be: Fix the wifi log flooding

+++ drivers/net/wireless/rockchip_wlan/rtl8852be/phl/phl_debug.h
@@ -61,7 +61,7 @@ enum {
 #define COMP_PHL_TWT BIT19
 
 extern u32 phl_log_components;
-extern u8 phl_log_level;
+extern uint phl_log_level;
 #define DEBUG_MAX_NAME_LEN 50
 
 struct dbg_alloc_buf {
